+++ src/library/UniformResourceIdentifier/Data/Image/index.ts
@@ -5,7 +5,7 @@ import type {
   DataURI_EncodingIdentifier,
 } from '../EncodingIdentifier';
 
-import DataURI from '../index.ts';
+import DataURI from '../definition.ts';
 
 import type {
   ImageURI_Format,
