+++ src/library/UniformResourceIdentifier/Data/definition.ts
